backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain>=0.3.7
langchain-openai>=0.2.8
langchain-groq>=0.2.1
langchain-google-genai>=2.1.12
python-dotenv==1.0.1
pydantic==2.9.2
slowapi==0.1.9
httpx==0.27.2

# Database
sqlalchemy==2.0.23
asyncpg==0.30.0
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.6
email-validator>=2.0.0

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
